Update main.tf
Add null_resource to trigger data regenerate for every run. Normally one would not need to regenerate data for every terraform run when interacting manually, this takes a long time. The depends_on line can be commented out in that case. If this code is run in a pipeline on a schedule once or twice a week, leave it uncommented.